trading_bot/requirements.txt
python-dotenv
# add real exchange client libs as needed, e.g. python-binance
